Prototype screwless plate for NZ/AU 118mmX74mm
FreeCad prototypes and exported AMF for each of the 2.4", 2.8" and
3.2" nextion displays.
Using heatset inserts that are M3 thread, 4.2mm OD

At present this design "clips" onto a mythical plastic plate which
will sit flat on the wall, screwed onto the flushbox / gib anchor
(HPM 712 or similar), this approach allows a screen larger than
2.4" to be used. And allows tool-less removal of the screen which
might be a downside for protecting end-users from 240VAC

This design requires the "grove" connector on the back to be removed,
much like upstream HASP enclosures do. The 6mm protrusion is just
too much to fit cleanly on the wall.