Icon-only version of the Weather Condition
Forked the weather-condition-with-icon blueprint, which
has a right-aligned text description of the weather and a super-magic
overstrike icon left-aligned of the weather condition.

Removed the text field entirely, so we get the same icons for the
same conditions, but centred.
Suitable for use on small buttons to convey at a glance approximately
what the weather is doing