Blueprint to control Tasmota/Tuya Curtain
I have curtain motors running Tuya converted to Tasmota
These advertise as lights at the moment with curtain position
presented as brightness.

This automation lets me set the curtain to any position, with
a dimmer for infinite variation, and three buttons, notionally
"open", "close", "halfway"; but text and values can be customised.

But I think I depend on a custom HMI, so this automation is not much
use to people who aren't me. (the custom HMI is over in the HMI
directory, tho)